Moved presentation classes to production
