+++ EssentialFeed/EssentialFeed/ImageComments Presentation/ImageCommentErrorView.swift
@@ -1,0 +1,11 @@
+//
+//  ImageCommentErrorView.swift
+//  EssentialFeed
+//
+//  Created by Eric Garlock on 3/10/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+public protocol ImageCommentErrorView {
+	func display(_ viewModel: ImageCommentErrorViewModel)
+}
