+++ EssentialFeed/EssentialFeed/ImageComments Presentation/ImageCommentLoadingView.swift
@@ -1,0 +1,11 @@
+//
+//  ImageCommentLoadingView.swift
+//  EssentialFeed
+//
+//  Created by Eric Garlock on 3/10/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+public protocol ImageCommentLoadingView {
+	func display(_ viewModel: ImageCommentLoadingViewModel)
+}
