+++ EssentialFeed/EssentialFeed/ImageComments Presentation/ImageCommentView.swift
@@ -1,0 +1,11 @@
+//
+//  ImageCommentView.swift
+//  EssentialFeed
+//
+//  Created by Eric Garlock on 3/10/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+public protocol ImageCommentView {
+	func display(_ viewModel: ImageCommentsViewModel)
+}
