+++ EssentialFeed/EssentialFeed/ImageComments Presentation/ImageCommentsViewModel.swift
@@ -1,0 +1,11 @@
+//
+//  ImageCommentsViewModel.swift
+//  EssentialFeed
+//
+//  Created by Eric Garlock on 3/10/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+public struct ImageCommentsViewModel {
+	public let comments: [ImageCommentViewModel]
+}
