+++ EssentialFeed/EssentialFeed/ImageComments Presentation/pt-BR.lproj/ImageComments.strings
