+++ EssentialFeed/EssentialFeedTests/ImageComments Presentation/ImageCommentLocalizationTests.swift
@@ -7,6 +7,7 @@
 //
 
 import XCTest
+import EssentialFeed
 
 class ImageCommentLocalizationTests: XCTestCase {
